implemented logic to have contest button in separate component
