+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/contests/contest-button/ContestButton.tsx
@@ -1,0 +1,46 @@
+import { useNavigate } from 'react-router-dom';
+
+import { useAppSelector } from '../../../redux/store';
+import { generateContestBtnUrlString } from '../../../utils/constants';
+import Button, { ButtonSize, ButtonState } from '../../guidelines/buttons/Button';
+
+interface IContestButtonProps {
+    isCompete: boolean;
+    isDisabled: boolean;
+    id: number;
+}
+
+const COMPETE_STRING = 'COMPETE';
+const PRACTICE_STRING = 'PRACTICE';
+
+const ContestButton = (props: IContestButtonProps) => {
+    const { isCompete, isDisabled, id } = props;
+
+    const { isLoggedIn } = useAppSelector((state) => state.authorization);
+    const navigate = useNavigate();
+
+    const btnText = isCompete
+        ? COMPETE_STRING
+        : PRACTICE_STRING;
+    const btnNavigateUrl = generateContestBtnUrlString(isCompete, id);
+
+    return (
+        <Button
+          text={btnText}
+          state={isDisabled
+              ? ButtonState.disabled
+              : ButtonState.enabled}
+          size={ButtonSize.small}
+          isCompete={isCompete}
+          onClick={() => {
+              if (!isLoggedIn) {
+                  navigate('/login');
+                  return;
+              }
+              navigate(btnNavigateUrl!);
+          }}
+        />
+    );
+};
+
+export default ContestButton;
