+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/utils/constants.ts
@@ -2,9 +2,6 @@ const DEFAULT_ITEMS_PER_PAGE = 10;
 
 const DEFAULT_ROWS_PER_PAGE = [ 10, 25, 50, 100 ];
 
-const COMPETE_STRING = 'COMPETE';
-const PRACTICE_STRING = 'PRACTICE';
-
 // eslint-disable-next-line consistent-return
 const generateContestBtnUrlString = (isCompete: boolean, id?: number) => {
     if (id) {
@@ -18,7 +15,5 @@ const generateContestBtnUrlString = (isCompete: boolean, id?: number) => {
 export {
     DEFAULT_ROWS_PER_PAGE,
     DEFAULT_ITEMS_PER_PAGE,
-    PRACTICE_STRING,
-    COMPETE_STRING,
     generateContestBtnUrlString,
 };
